+++ README_DIR/CHANGELOG.md
@@ -6,6 +6,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## Unreleased Pet Food Calculator
 
+## [2.1.6-beta] - 2025-2-25
+
+## Added
+
+- Feature: Create file for MongoDB integration
+
+## Changed
+
+- UI: Adjust label size for login and register
+  
+
 ## [2.1.5-beta] - 2025-2-25
 
 ## Added